% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blblm.R
\name{blblm}
\alias{blblm}
\title{linear regression with mini bag bootstrap}
\usage{
blblm(formula, data, m = 10, B = 5000, num_worker = 1)
}
\arguments{
\item{formula}{formula}

\item{data}{dataset}

\item{m}{how many times we split the dataset}

\item{B}{number of bootstrap performed}

\item{num_worker}{number of workers when we run parallel}
}
\value{
blblm class object
}
\description{
linear regression with mini bag bootstrap
}
