% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blblm.R
\name{coef.blblm}
\alias{coef.blblm}
\title{get coefficients from blblm object}
\usage{
\method{coef}{blblm}(object, ...)
}
\arguments{
\item{object}{blblm output}

\item{...}{other parameters}
}
\description{
get coefficients from blblm object
}
