% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blblm.R
\name{confint.blblm}
\alias{confint.blblm}
\title{get confidence interval of blblm object}
\usage{
\method{confint}{blblm}(object, parm = NULL, level = 0.95, ...)
}
\arguments{
\item{object}{blblm output}

\item{parm}{parameters}

\item{level}{confidence level, default 0.95}

\item{...}{other parameters}
}
\description{
get confidence interval of blblm object
}
