% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blblm.R
\name{lm1}
\alias{lm1}
\title{estimate the regression estimates based on given the number of repetitions}
\usage{
lm1(formula, data, freqs)
}
\arguments{
\item{formula}{formula}

\item{data}{dataset}

\item{freqs}{frequents how many times observations are repeated}
}
\description{
estimate the regression estimates based on given the number of repetitions
}
