% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blblm.R
\name{predict.blblm}
\alias{predict.blblm}
\title{prediciton on blblm object}
\usage{
\method{predict}{blblm}(object, new_data, confidence = FALSE, level = 0.95, ...)
}
\arguments{
\item{object}{blblm object}

\item{new_data}{dataset we want to predict}

\item{confidence}{need confidence interval or not}

\item{level}{confidence level}

\item{...}{other parameters}
}
\description{
prediciton on blblm object
}
