% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blblm.R
\name{sigma.blblm}
\alias{sigma.blblm}
\title{print method on blblm}
\usage{
\method{sigma}{blblm}(object, confidence = FALSE, level = 0.95, ...)
}
\arguments{
\item{object}{blblm object}

\item{confidence}{need confidence interval or not, default is FALSE}

\item{level}{level of the confidence intervals, default 0.95}

\item{...}{other parameters}
}
\description{
print method on blblm
}
